+++ src/lib/mina.ts
@@ -15,6 +15,7 @@ import {
   Actions,
   Events,
   dummySignature,
+  AccountUpdateLayout,
 } from './account_update.js';
 import * as Fetch from './fetch.js';
 import { assertPreconditionInvariants, NetworkValue } from './precondition.js';
@@ -181,7 +182,7 @@ function createTransaction(
 
   let transactionId = currentTransaction.enter({
     sender,
-    accountUpdates: [],
+    layout: new AccountUpdateLayout(),
     fetchMode,
     isFinalRunOutsideCircuit,
     numberOfRuns,
@@ -203,9 +204,7 @@ function createTransaction(
             f();
             Provable.asProver(() => {
               let tx = currentTransaction.get();
-              tx.accountUpdates = CallForest.map(tx.accountUpdates, (a) =>
-                toConstant(AccountUpdate, a)
-              );
+              tx.layout.toConstantInPlace();
             });
           });
         } else {
@@ -221,10 +220,10 @@ function createTransaction(
     currentTransaction.leave(transactionId);
     throw err;
   }
-  let accountUpdates = currentTransaction.get().accountUpdates;
-  // TODO: I'll be back
-  // CallForest.addCallers(accountUpdates);
-  accountUpdates = CallForest.toFlatList(accountUpdates);
+
+  let accountUpdates = currentTransaction
+    .get()
+    .layout.toFlatList({ mutate: true });
 
   try {
     // check that on-chain values weren't used without setting a precondition
@@ -428,9 +427,11 @@ function LocalBlockchain({
 
         // TODO: verify account update even if the account doesn't exist yet, using a default initial account
         if (account !== undefined) {
+          let publicInput = update.toPublicInput(txn.transaction);
           await verifyAccountUpdate(
             account,
             update,
+            publicInput,
             commitments,
             this.proofsEnabled,
             this.getNetworkId()
@@ -1163,6 +1164,7 @@ function defaultNetworkState(): NetworkValue {
 async function verifyAccountUpdate(
   account: Account,
   accountUpdate: AccountUpdate,
+  publicInput: ZkappPublicInput,
   transactionCommitments: { commitment: bigint; fullCommitment: bigint },
   proofsEnabled: boolean,
   networkId: NetworkId
@@ -1244,7 +1246,6 @@ async function verifyAccountUpdate(
 
   if (accountUpdate.authorization.proof && proofsEnabled) {
     try {
-      let publicInput = accountUpdate.toPublicInput();
       let publicInputFields = ZkappPublicInput.toFields(publicInput);
 
       let proof: JsonProof = {