+++ src/lib/mina/transaction-context.ts
@@ -1,4 +1,4 @@
-import type { AccountUpdate } from '../account_update.js';
+import type { AccountUpdateLayout } from '../account_update.js';
 import type { PublicKey } from '../signature.js';
 import { Context } from '../global-context.js';
 
@@ -7,7 +7,7 @@ export { currentTransaction, CurrentTransaction, FetchMode };
 type FetchMode = 'fetch' | 'cached' | 'test';
 type CurrentTransaction = {
   sender?: PublicKey;
-  accountUpdates: AccountUpdate[];
+  layout: AccountUpdateLayout;
   fetchMode: FetchMode;
   isFinalRunOutsideCircuit: boolean;
   numberOfRuns: 0 | 1 | undefined;